ai-site-monitorV2_package/requirements.txt
aiohttp
beautifulsoup4
pyyaml
python-dotenv
google-generativeai
